+++ app/src/main/java/com/hongmyeoun/goldcalc/view/home/topbar/progressText/ProgressText.kt
@@ -25,7 +25,7 @@ import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.hongmyeoun.goldcalc.model.common.ImageReturn
 import com.hongmyeoun.goldcalc.model.common.formatWithCommas
-import com.hongmyeoun.goldcalc.view.characterDetail.normalTextStyle
+import com.hongmyeoun.goldcalc.view.profile.normalTextStyle
 import com.hongmyeoun.goldcalc.viewModel.main.CharacterListVM
 
 @OptIn(ExperimentalGlideComposeApi::class)
