+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/card/Simple.kt
@@ -1,0 +1,33 @@
+package com.hongmyeoun.goldcalc.view.profile.content.card
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.hongmyeoun.goldcalc.model.searchedInfo.card.Cards
+import com.hongmyeoun.goldcalc.viewModel.charDetail.CardDetailVM
+
+@Composable
+fun Simple(
+    cardList: List<Cards>,
+    viewModel: CardDetailVM
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(4.dp),
+        horizontalArrangement = Arrangement.SpaceEvenly
+    ) {
+        cardList.forEach { card ->
+            CardImage(
+                grade = card.grade,
+                icon = card.icon,
+                awakeCount = card.awakeCount,
+                viewModel = viewModel
+            )
+        }
+    }
+}
