+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/equipment/Equipments.kt
@@ -1,4 +1,4 @@
-package com.hongmyeoun.goldcalc.view.characterDetail.equipment
+package com.hongmyeoun.goldcalc.view.profile.content.equipment
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -15,6 +15,8 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -29,9 +31,23 @@ import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.CharacterEquipment
 import com.hongmyeoun.goldcalc.ui.theme.BlackTransBG
 import com.hongmyeoun.goldcalc.ui.theme.GreenQual
 import com.hongmyeoun.goldcalc.ui.theme.ImageBG
-import com.hongmyeoun.goldcalc.view.characterDetail.TextChip
-import com.hongmyeoun.goldcalc.view.characterDetail.normalTextStyle
+import com.hongmyeoun.goldcalc.view.profile.normalTextStyle
+import com.hongmyeoun.goldcalc.view.common.TextChip
+import com.hongmyeoun.goldcalc.viewModel.charDetail.CharDetailVM
 import com.hongmyeoun.goldcalc.viewModel.charDetail.EquipmentDetailVM
+
+@Composable
+fun Equipment(
+    viewModel: CharDetailVM
+) {
+    // 장비
+    val equipment by viewModel.equipments.collectAsState()
+
+    equipment?.let { equipmentList ->
+        val equipmentVM = EquipmentDetailVM(equipmentList)
+        EquipmentDetailUI(equipmentList, equipmentVM)
+    }
+}
 
 @Composable
 fun EquipmentDetails(
