+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/gem/Detail.kt
@@ -1,12 +1,9 @@
-package com.hongmyeoun.goldcalc.view.characterDetail
+package com.hongmyeoun.goldcalc.view.profile.content.gem
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -17,8 +14,6 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -27,93 +22,36 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.hongmyeoun.goldcalc.model.searchedInfo.gem.Gem
 import com.hongmyeoun.goldcalc.ui.theme.BlackTransBG
 import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
-import com.hongmyeoun.goldcalc.ui.theme.LightGrayTransBG
 import com.hongmyeoun.goldcalc.ui.theme.RelicColor
-import com.hongmyeoun.goldcalc.viewModel.charDetail.GemDetailVM
+import com.hongmyeoun.goldcalc.view.profile.normalTextStyle
+import com.hongmyeoun.goldcalc.view.common.TextChip
 
 @Composable
-@OptIn(ExperimentalLayoutApi::class)
-fun GemDetailUI(gemList: List<Gem>, viewModel: GemDetailVM = viewModel()) {
-    val isDetail by viewModel.isDetail.collectAsState()
-
-    Column(
-        modifier = Modifier
-            .background(LightGrayTransBG, RoundedCornerShape(8.dp))
-            .clip(RoundedCornerShape(8.dp))
-            .padding(8.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable { viewModel.onDetailClicked() }
-        ) {
-            Text(
-                text = "보석",
-                style = titleTextStyle()
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-
-        }
-        if (isDetail) {
-            Row {
-                GemDetail(
-                    modifier = Modifier.weight(1f),
-                    effectType = "피해",
-                    gemList = gemList
-                )
-                GemDetail(
-                    modifier = Modifier.weight(1f),
-                    effectType = "재사용 대기시간",
-                    gemList = gemList
-                )
-            }
-        } else {
-            val (annihilation, crimsonFlame) = viewModel.countAnnihilationGem(gemList)
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-            ) {
-                TextChip(text = "멸화 x$annihilation")
-                Spacer(modifier = Modifier.width(8.dp))
-                TextChip(text = "홍염 x$crimsonFlame")
-            }
-            Spacer(modifier = Modifier.height(4.dp))
-
-            Row {
-                val (annMaxItemCount, criMaxItemCount) = viewModel.calcMaxItemsInEachRow(annihilation, crimsonFlame)
-                Column {
-                    FlowRow(
-                        maxItemsInEachRow = annMaxItemCount
-                    ) {
-                        gemList.filter { it.type == "멸화" }.forEach {
-                            GemSimple(it, viewModel)
-                        }
-                    }
-                }
-                Column {
-                    FlowRow(
-                        maxItemsInEachRow = criMaxItemCount
-                    ) {
-                        gemList.filter { it.type == "홍염" }.forEach {
-                            GemSimple(it, viewModel)
-                        }
-                    }
-                }
-            }
-        }
+fun Detail(
+    gemList: List<Gem>,
+) {
+    Row {
+        GemDetail(
+            modifier = Modifier.weight(1f),
+            effectType = "피해",
+            gemList = gemList
+        )
+        GemDetail(
+            modifier = Modifier.weight(1f),
+            effectType = "재사용 대기시간",
+            gemList = gemList
+        )
     }
-    Spacer(modifier = Modifier.height(8.dp))
 }
 
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
-fun GemDetail(
+private fun GemDetail(
     modifier: Modifier,
     effectType: String,
     gemList: List<Gem>
@@ -186,32 +124,5 @@ fun GemDetail(
                 }
             }
         }
-    }
-}
-
-@Composable
-@OptIn(ExperimentalGlideComposeApi::class)
-private fun GemSimple(it: Gem, viewModel: GemDetailVM) {
-    Column(
-        modifier = Modifier
-            .padding(4.dp)
-            .background(
-                color = LightGrayBG,
-                shape = RoundedCornerShape(4.dp)
-            ),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        GlideImage(
-            modifier = Modifier
-                .size(44.dp)
-                .background(
-                    brush = viewModel.getItemBG(it.grade),
-                    shape = RoundedCornerShape(topStart = 4.dp, topEnd = 4.dp)
-                )
-                .padding(4.dp),
-            model = it.gemIcon,
-            contentDescription = "보석",
-        )
-        Text(text = "${it.level}", fontSize = 10.sp, color = Color.White)
     }
 }
