+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/gem/GemView.kt
@@ -1,0 +1,62 @@
+package com.hongmyeoun.goldcalc.view.profile.content.gem
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.hongmyeoun.goldcalc.model.searchedInfo.gem.Gem
+import com.hongmyeoun.goldcalc.ui.theme.LightGrayTransBG
+import com.hongmyeoun.goldcalc.view.profile.Title
+import com.hongmyeoun.goldcalc.viewModel.charDetail.CharDetailVM
+import com.hongmyeoun.goldcalc.viewModel.charDetail.GemDetailVM
+
+@Composable
+fun Gem(
+    viewModel: CharDetailVM
+) {
+    // 보석
+    val gems by viewModel.gems.collectAsState()
+
+    gems?.let { gemList ->
+        GemView(gemList)
+    }
+}
+
+@Composable
+fun GemView(
+    gemList: List<Gem>,
+    viewModel: GemDetailVM = viewModel()
+) {
+    val isDetail by viewModel.isDetail.collectAsState()
+
+    Column(
+        modifier = Modifier
+            .background(LightGrayTransBG, RoundedCornerShape(8.dp))
+            .clip(RoundedCornerShape(8.dp))
+            .padding(8.dp)
+    ) {
+        Title(
+            title = "보석",
+            onClick = { viewModel.onDetailClicked() }
+        )
+
+        if (isDetail) {
+            Detail(gemList = gemList)
+        } else {
+            Simple(
+                gemList = gemList,
+                viewModel = viewModel
+            )
+        }
+    }
+    Spacer(modifier = Modifier.height(8.dp))
+}
