+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/skill/Detail.kt
@@ -1,4 +1,4 @@
-package com.hongmyeoun.goldcalc.view.characterDetail
+package com.hongmyeoun.goldcalc.view.profile.content.skill
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -17,8 +17,6 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -27,73 +25,21 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
-import com.hongmyeoun.goldcalc.model.lostArkApi.CharacterDetail
 import com.hongmyeoun.goldcalc.model.searchedInfo.gem.Gem
 import com.hongmyeoun.goldcalc.model.searchedInfo.skills.Skills
 import com.hongmyeoun.goldcalc.model.searchedInfo.skills.Tripods
-import com.hongmyeoun.goldcalc.ui.theme.BlackTransBG
 import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
 import com.hongmyeoun.goldcalc.ui.theme.LightGrayTransBG
 import com.hongmyeoun.goldcalc.ui.theme.PurpleQual
-import com.hongmyeoun.goldcalc.view.common.noRippleClickable
+import com.hongmyeoun.goldcalc.view.profile.normalTextStyle
+import com.hongmyeoun.goldcalc.view.profile.titleTextStyle
+import com.hongmyeoun.goldcalc.view.common.TextChip
 import com.hongmyeoun.goldcalc.viewModel.charDetail.SkillDetailVM
 
 @Composable
-fun SkillDetailUI(
-    characterDetail: CharacterDetail?,
-    skills: List<Skills>,
-    gemList: List<Gem>?,
-    viewModel: SkillDetailVM = viewModel()
-) {
-    val isDetail by viewModel.isDetail.collectAsState()
-
-    Column(
-        modifier = Modifier
-            .background(LightGrayTransBG, RoundedCornerShape(8.dp))
-            .clip(RoundedCornerShape(8.dp))
-            .noRippleClickable { viewModel.onDetailClicked() }
-            .padding(8.dp)
-    ) {
-        Text(
-            text = "스킬",
-            style = titleTextStyle()
-        )
-        Spacer(modifier = Modifier.height(4.dp))
-
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-        ) {
-            characterDetail?.let { characterDetail ->
-                TextChip(text = "스킬 포인트 ${characterDetail.usingSkillPoint}/${characterDetail.totalSkillPoint}")
-                Spacer(modifier = Modifier.width(8.dp))
-            }
-            val (fiveLevel, fourLevel) = viewModel.tripodLevel54(skills)
-
-            if (fiveLevel > 0) {
-                TextChip(text = "Lv.5 x$fiveLevel")
-                Spacer(modifier = Modifier.width(8.dp))
-            }
-            if (fourLevel > 0) {
-                TextChip(text = "Lv.4 x$fourLevel")
-            }
-        }
-        Spacer(modifier = Modifier.height(4.dp))
-
-        if (isDetail) {
-            SkillDetail(skills, gemList, viewModel)
-        } else {
-            SkillSimple(skills, gemList, viewModel)
-        }
-
-    }
-}
-
-@Composable
-private fun SkillDetail(
+fun Detail(
     skills: List<Skills>,
     gemList: List<Gem>?,
     viewModel: SkillDetailVM
@@ -119,14 +65,14 @@ private fun SkillDetail(
         Spacer(modifier = Modifier.height(4.dp))
 
         skills.forEach {
-            SkillDetailRow(it, gemList, viewModel)
+            DetailRow(it, gemList, viewModel)
             Spacer(modifier = Modifier.height(8.dp))
         }
     }
 }
 
 @Composable
-private fun SkillDetailRow(
+private fun DetailRow(
     it: Skills,
     gemList: List<Gem>?,
     viewModel: SkillDetailVM
@@ -146,79 +92,14 @@ private fun SkillDetailRow(
         ) {
             SkillIcon(it)
             Spacer(modifier = Modifier.width(4.dp))
-            SkillTripods(it, viewModel)
+
+            Tripods(it, viewModel)
         }
         Spacer(modifier = Modifier.height(8.dp))
 
         gemList?.let { gemList ->
             if (it.rune != null || it.gem) {
-                SkillRuneAndGem(it, gemList, viewModel)
-            }
-        }
-    }
-}
-
-@Composable
-@OptIn(ExperimentalGlideComposeApi::class)
-private fun SkillIcon(it: Skills, isSimple: Boolean = false) {
-    val iconSize = if (isSimple) 44.dp else 54.dp
-    Box {
-        Column {
-            Row {
-                Box {
-                    GlideImage(
-                        modifier = Modifier
-                            .size(iconSize)
-                            .border(
-                                width = 1.dp,
-                                color = Color.White,
-                                shape = RoundedCornerShape(10.dp)
-                            )
-                            .clip(RoundedCornerShape(10.dp)),
-                        model = it.icon,
-                        contentScale = ContentScale.Crop,
-                        contentDescription = "스킬 아이콘"
-                    )
-
-                    if (isSimple) {
-                        Column(modifier = Modifier.padding(2.dp)) {
-                            var tripodSlots = ""
-                            var tripodLevels = ""
-                            it.tripods.forEach { tripods ->
-                                if (tripods.isSelected) {
-                                    tripodSlots += tripods.slot.toString()
-                                    tripodLevels += tripods.level.toString()
-                                }
-                            }
-                            TextChip(
-                                text = tripodSlots,
-                                borderless = true,
-                                customTextSize = 8.sp,
-                                customBGColor = BlackTransBG,
-                                customPadding = Modifier.padding(start = 2.dp, end = 2.dp)
-                            )
-                            Spacer(modifier = Modifier.height(1.dp))
-                            TextChip(
-                                text = tripodLevels,
-                                borderless = true,
-                                customTextSize = 8.sp,
-                                customBGColor = BlackTransBG,
-                                customPadding = Modifier.padding(start = 2.dp, end = 2.dp)
-                            )
-                        }
-                    }
-                }
-                if (isSimple) {
-                    Spacer(modifier = Modifier.width(10.dp))
-                }
-            }
-            if (isSimple) {
-                Spacer(modifier = Modifier.height(6.dp))
-            }
-        }
-        if (isSimple) {
-            Box(modifier = Modifier.align(Alignment.BottomEnd)) {
-                TextChip(text = "${it.level}")
+                RuneAndGem(it, gemList, viewModel)
             }
         }
     }
@@ -226,7 +107,7 @@ private fun SkillIcon(it: Skills, isSimple: Boolean = false) {
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
-private fun SkillTripods(it: Skills, viewModel: SkillDetailVM) {
+private fun Tripods(it: Skills, viewModel: SkillDetailVM) {
     Column(horizontalAlignment = Alignment.Start) {
         Row(
             modifier = Modifier.fillMaxWidth(),
@@ -294,13 +175,16 @@ private fun TripodInfo(
     viewModel: SkillDetailVM,
     selectedIndex: Int
 ) {
-    Row(verticalAlignment = Alignment.CenterVertically) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically
+    ) {
         TextChip(
             text = "${tripod.level}",
             borderless = true,
             customBGColor = viewModel.getIndexColor(selectedIndex)
         )
         Spacer(modifier = Modifier.width(4.dp))
+
         Text(
             text = tripod.name,
             style = normalTextStyle(color = viewModel.getIndexColor(selectedIndex), fontSize = 12.sp)
@@ -309,10 +193,9 @@ private fun TripodInfo(
     }
 }
 
-
 @Composable
 @OptIn(ExperimentalGlideComposeApi::class)
-private fun SkillRuneAndGem(
+private fun RuneAndGem(
     it: Skills,
     gemList: List<Gem>,
     viewModel: SkillDetailVM
@@ -415,103 +298,6 @@ private fun SkillRuneAndGem(
                     }
 
                 }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun SkillSimple(skills: List<Skills>, gemList: List<Gem>?, viewModel: SkillDetailVM) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-    ) {
-        FlowRow(
-            maxItemsInEachRow = 2,
-            horizontalArrangement = Arrangement.SpaceEvenly
-        ) {
-            skills.forEach {
-                if (it.level > 3) {
-                    Row(
-                        modifier = Modifier
-                            .padding(4.dp)
-                            .background(
-                                color = LightGrayBG,
-                                shape = RoundedCornerShape(8.dp)
-                            )
-                            .padding(4.dp)
-                            .weight(1f),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        SkillIcon(it, true)
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Column(
-                            horizontalAlignment = Alignment.Start
-                        ) {
-                            SkillSimpleTripods(it)
-                            Spacer(modifier = Modifier.height(4.dp))
-                            gemList?.let { gemList ->
-                                SkillSimpleRuneAndGem(it, gemList, viewModel)
-                            }
-                        }
-                    }
-                }
-                if (it == skills.last()) {
-                    if (skills.size % 2 == 1) {
-                        Spacer(
-                            modifier = Modifier
-                                .width(1.dp)
-                                .weight(1f)
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun SkillSimpleTripods(it: Skills) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = it.name,
-            style = titleBoldWhite12()
-        )
-    }
-}
-
-@Composable
-private fun SkillSimpleRuneAndGem(
-    it: Skills,
-    gemList: List<Gem>,
-    viewModel: SkillDetailVM
-) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        it.rune?.let { rune ->
-            TextChip(
-                text = rune.name,
-                customBGColor = viewModel.getGradeBG(rune.grade),
-                borderless = true
-            )
-            Spacer(modifier = Modifier.width(4.dp))
-        }
-        if (it.gem) {
-            val equipGemList = gemList.filter { gem -> gem.skill == it.name }
-            equipGemList.forEach { gem ->
-                val gemInfo = viewModel.gemInfo(gem)
-
-                val padding = if (gemInfo.length >= 3 && equipGemList.size > 1) {
-                    Modifier.padding(start = 4.dp, end = 4.dp, top = 2.dp, bottom = 2.dp)
-                } else {
-                    Modifier.padding(start = 6.dp, end = 6.dp, top = 2.dp, bottom = 2.dp)
-                }
-                TextChip(text = gemInfo, borderless = true, customPadding = padding)
-                Spacer(modifier = Modifier.width(4.dp))
             }
         }
     }
