+++ app/src/main/java/com/hongmyeoun/goldcalc/view/search/content/SearchResult.kt
@@ -32,7 +32,7 @@ import com.hongmyeoun.goldcalc.model.lostArkApi.CharacterInfo
 import com.hongmyeoun.goldcalc.model.lostArkApi.CharacterResourceMapper
 import com.hongmyeoun.goldcalc.navigation.Screen
 import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
-import com.hongmyeoun.goldcalc.view.characterDetail.titleTextStyle
+import com.hongmyeoun.goldcalc.view.profile.titleTextStyle
 
 @Composable
 @OptIn(ExperimentalFoundationApi::class)
